cmd/cored: introduce no_mockhsm build tag
By default, the mockhsm will be enabled. To
disable its usage, the cored binary should
be built with the `no_mockhsm` flag.
