+++ cmd/cored/no_mockhsm.go
@@ -1,0 +1,23 @@
+//+build no_mockhsm
+
+package main
+
+import (
+	"chain/core"
+	"chain/core/blocksigner"
+	"chain/core/config"
+	"chain/database/pg"
+	"errors"
+)
+
+func init() {
+	config.BuildConfig.MockHSM = false
+}
+
+func devEnableMockHSM(pg.DB) []core.RunOption {
+	return nil
+}
+
+func devHSM(pg.DB) (blocksigner.Signer, error) {
+	return nil, errors.New("cannot use mockhsm in production, must configure block hsm url")
+}
