+++ cmd/cored/prod.go
@@ -18,11 +18,3 @@ func resetInDevIfRequested(db pg.DB) {}
 func authLoopbackInDev(req *http.Request) bool {
 	return false
 }
-
-func devEnableMockHSM(_ pg.DB) []core.RunOption {
-	return nil
-}
-
-func devHSM(_ pg.DB) (blocksigner.Signer, error) {
-	return nil, errors.New("cannot use mockhsm in production, must configure block hsm url")
-}
